add localizations
