+++ Core/Core/SwiftGen/Strings.swift
@@ -124,10 +124,14 @@ public enum CoreLocalization {
     public static let search = CoreLocalization.tr("Localizable", "PICKER.SEARCH", fallback: "Search")
   }
   public enum Review {
+    /// What could have been better?
+    public static let better = CoreLocalization.tr("Localizable", "REVIEW.BETTER", fallback: "What could have been better?")
     /// We’re sorry to hear your learning experience has had some issues. We appreciate all feedback.
     public static let feedbackDescription = CoreLocalization.tr("Localizable", "REVIEW.FEEDBACK_DESCRIPTION", fallback: "We’re sorry to hear your learning experience has had some issues. We appreciate all feedback.")
     /// Leave Us Feedback
     public static let feedbackTitle = CoreLocalization.tr("Localizable", "REVIEW.FEEDBACK_TITLE", fallback: "Leave Us Feedback")
+    /// Not now
+    public static let notNow = CoreLocalization.tr("Localizable", "REVIEW.NOT_NOW", fallback: "Not now")
     /// We received your feedback and will use it to help improve your learning experience going forward. Thank you for sharing!
     public static let thanksForFeedbackDescription = CoreLocalization.tr("Localizable", "REVIEW.THANKS_FOR_FEEDBACK_DESCRIPTION", fallback: "We received your feedback and will use it to help improve your learning experience going forward. Thank you for sharing!")
     /// Thank You
@@ -140,6 +144,18 @@ public enum CoreLocalization {
     public static let voteDescription = CoreLocalization.tr("Localizable", "REVIEW.VOTE_DESCRIPTION", fallback: "Your feedback matters to us. Would you take a moment to rate the app by tapping a star below? Thanks for your support!")
     /// Enjoying Open edX?
     public static let voteTitle = CoreLocalization.tr("Localizable", "REVIEW.VOTE_TITLE", fallback: "Enjoying Open edX?")
+    public enum Button {
+      /// Rate Us
+      public static let rateUs = CoreLocalization.tr("Localizable", "REVIEW.BUTTON.RATE_US", fallback: "Rate Us")
+      /// Share Feedback
+      public static let shareFeedback = CoreLocalization.tr("Localizable", "REVIEW.BUTTON.SHARE_FEEDBACK", fallback: "Share Feedback")
+      /// Submit
+      public static let submit = CoreLocalization.tr("Localizable", "REVIEW.BUTTON.SUBMIT", fallback: "Submit")
+    }
+    public enum Email {
+      /// Select email client:
+      public static let title = CoreLocalization.tr("Localizable", "REVIEW.EMAIL.TITLE", fallback: "Select email client:")
+    }
   }
   public enum View {
     public enum Snackbar {
